Delete warehouse proc related code
