+++ fe/fe-core/src/main/java/com/starrocks/common/proc/ProcService.java
@@ -37,7 +37,6 @@ package com.starrocks.common.proc;
 import com.google.common.base.Strings;
 import com.starrocks.common.AnalysisException;
 import com.starrocks.server.GlobalStateMgr;
-import com.starrocks.warehouse.WarehouseProcDir;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -70,7 +69,6 @@ public final class ProcService {
         root.register("colocation_group", new ColocationGroupProcDir());
         root.register("catalog", GlobalStateMgr.getCurrentState().getCatalogMgr().getProcNode());
         root.register("compactions", new CompactionsProcNode());
-        root.register("warehouses", new WarehouseProcDir(GlobalStateMgr.getCurrentState().getWarehouseMgr()));
     }
 
     // Get the corresponding PROC Node by the specified path
