+++ fe/fe-core/src/main/java/com/starrocks/server/WarehouseManager.java
@@ -16,21 +16,16 @@ package com.starrocks.server;
 
 import com.google.common.collect.ImmutableMap;
 import com.staros.util.LockCloseable;
-import com.starrocks.common.DdlException;
 import com.starrocks.common.io.Text;
 import com.starrocks.common.io.Writable;
 import com.starrocks.persist.gson.GsonUtils;
 import com.starrocks.system.ComputeNode;
-import com.starrocks.warehouse.LocalWarehouse;
+import com.starrocks.warehouse.DefaultWarehouse;
 import com.starrocks.warehouse.Warehouse;
-import com.starrocks.warehouse.WarehouseProcDir;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.io.DataInputStream;
 import java.io.DataOutput;
-import java.io.DataOutputStream;
-import java.io.EOFException;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.HashMap;
@@ -46,6 +41,8 @@ public class WarehouseManager implements Writable {
 
     public static final long DEFAULT_WAREHOUSE_ID = 0L;
 
+    public static final long DEFAULT_CLUSTER_ID = 0L;
+
     private Map<Long, Warehouse> idToWh = new HashMap<>();
     private Map<String, Warehouse> nameToWh = new HashMap<>();
 
@@ -57,8 +54,8 @@ public class WarehouseManager implements Writable {
     public void initDefaultWarehouse() {
         // gen a default warehouse
         try (LockCloseable lock = new LockCloseable(rwLock.writeLock())) {
-            Warehouse wh = new LocalWarehouse(DEFAULT_WAREHOUSE_ID,
-                    DEFAULT_WAREHOUSE_NAME);
+            Warehouse wh = new DefaultWarehouse(DEFAULT_WAREHOUSE_ID,
+                    DEFAULT_WAREHOUSE_NAME, DEFAULT_CLUSTER_ID);
             nameToWh.put(wh.getName(), wh);
             idToWh.put(wh.getId(), wh);
             wh.setExist(true);
@@ -100,33 +97,6 @@ public class WarehouseManager implements Writable {
                 nodeId -> builder.put(nodeId,
                         GlobalStateMgr.getCurrentState().getNodeMgr().getClusterInfo().getBackendOrComputeNode(nodeId)));
         return builder.build();
-    }
-
-    // not persist anything thereafter, so checksum ^= 0
-    public long saveWarehouses(DataOutputStream out, long checksum) throws IOException {
-        checksum ^= 0;
-        write(out);
-        return checksum;
-    }
-
-    public long loadWarehouses(DataInputStream dis, long checksum) throws IOException, DdlException {
-        int warehouseCount = 0;
-        try {
-            String s = Text.readString(dis);
-            WarehouseManager data = GsonUtils.GSON.fromJson(s, WarehouseManager.class);
-            if (data != null && data.nameToWh != null) {
-                warehouseCount = data.nameToWh.size();
-            }
-            checksum ^= warehouseCount;
-            LOG.info("finished replaying WarehouseMgr from image");
-        } catch (EOFException e) {
-            LOG.info("no WarehouseMgr to replay.");
-        }
-        return checksum;
-    }
-
-    public List<List<String>> getWarehousesInfo() {
-        return new WarehouseProcDir(this).fetchResult().getRows();
     }
 
     @Override
