+++ fe/fe-core/src/main/java/com/starrocks/warehouse/Cluster.java
@@ -14,12 +14,10 @@
 
 package com.starrocks.warehouse;
 
-import com.google.common.collect.Lists;
 import com.google.gson.annotations.SerializedName;
 import com.starrocks.common.UserException;
 import com.starrocks.common.io.Text;
 import com.starrocks.common.io.Writable;
-import com.starrocks.common.proc.BaseProcResult;
 import com.starrocks.lake.StarOSAgent;
 import com.starrocks.persist.gson.GsonUtils;
 import com.starrocks.server.GlobalStateMgr;
@@ -27,7 +25,6 @@ import com.starrocks.server.RunMode;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
-import java.io.DataInput;
 import java.io.DataOutput;
 import java.io.IOException;
 import java.util.ArrayList;
@@ -54,21 +51,6 @@ public class Cluster implements Writable {
         return workerGroupId;
     }
 
-    public int getRunningSqls() {
-        return -1;
-    }
-    public int getPendingSqls() {
-        return -1;
-    }
-
-    public void getProcNodeData(BaseProcResult result) {
-        result.addRow(Lists.newArrayList(String.valueOf(this.getId()),
-                String.valueOf(this.getWorkerGroupId()),
-                String.valueOf(this.getComputeNodeIds()),
-                String.valueOf(this.getPendingSqls()),
-                String.valueOf(this.getRunningSqls())));
-    }
-
     public List<Long> getComputeNodeIds() {
         List<Long> nodeIds = new ArrayList<>();
         if (RunMode.isSharedNothingMode()) {
@@ -89,10 +71,4 @@ public class Cluster implements Writable {
         String json = GsonUtils.GSON.toJson(this);
         Text.writeString(out, json);
     }
-
-    public static Cluster read(DataInput in) throws IOException {
-        String json = Text.readString(in);
-        return GsonUtils.GSON.fromJson(json, Cluster.class);
-    }
-
 }
